Fixed some bugs

See merge request 1dv508/student/beryllium-vt22/cookbook-project/cookbook!200

+++ lib/models/ingredient/ingredient.dart
@@ -8,9 +8,9 @@ class Ingredient with ChangeNotifier {
   String name;
   double pricePerUnit;
   String unit;
-  int? amount;
+  double? amount;
   double? changedPrice;
-  int? changedAmount;
+  double? changedAmount;
   //final double amountFactor;
 
   Ingredient({
@@ -49,7 +49,7 @@ class Ingredient with ChangeNotifier {
     );
   }
 
-  void setAmount(int amountToAdd) {
+  void setAmount(double amountToAdd) {
     amount = amountToAdd;
     notifyListeners();
   }
